feat: add my solution
The results array is the array used for binary search, and the array ranges from 0 to the array's maximum value. This is because a value larger than the array element value cannot be subtracted. The difference from binary search is the condition of the if statements in the while statement.